Fix code structure
