--- src/modules/comments/commentHeader.js
@@ -1,9 +1,0 @@
-const commentHeader = () => {
-  let comments = `<div class="comments">
-    <h2 class="comments-heading">Comments</h2>
-    <h2 class="comments-num">(<span class="comments-count">0</span>)</h2>
-  </div>`;
-  return comments;
-};
-
-export default commentHeader;
