+++ src/modules/commentsCounter.js
