+++ src/modules/render.js
@@ -2,7 +2,8 @@ import '@fortawesome/fontawesome-free/js/fontawesome';
 import '@fortawesome/fontawesome-free/js/solid';
 import '@fortawesome/fontawesome-free/js/regular';
 import '@fortawesome/fontawesome-free/js/brands';
-import showPopup from './popup';
+import commentsCounter from './commentsCounter.js';
+import { postComment, getComment } from './involvement-api.js';
 
 const LIKES_COUNT = 5;
 
@@ -67,14 +68,84 @@ const renderMainList = (ebookList) => {
 
     const commentButton = document.createElement('button');
     commentButton.classList = 'contact-button';
+
     commentButton.addEventListener('click', async (e) => {
       e.preventDefault();
       popupSection.classList.toggle('hide');
-      showPopup(ebook);
+      const popup = document.querySelector('.popup');
+
+      const bookDetails = `<div class="item-details">
+      <div class="cross-icon">
+        <img src="https://cdn2.iconfinder.com/data/icons/flat-ui-icons-24-px/24/cross-24-512.png" alt="">
+      </div>
+      <img src="${ebook.artworkUrl100.replace('100x100', '300x300')}" class="book-img" alt="">
+      <div class="book-title text-center">
+        <h2>${ebook.trackName}</h2>
+      </div>
+      <div class="book-detail">
+        <p><b>Author:</b> ${ebook.artistName}</p>
+        <p><b>Price:</b> ${ebook.price}$</p>
+        <p><b>Description:</b> ${ebook.description}</p>
+      </div>
+      </div>`;
+      popup.insertAdjacentHTML('beforeend', bookDetails);
+
+      const crossIcon = document.querySelector('.cross-icon');
+      crossIcon.addEventListener('click', () => {
+      popupSection.classList.toggle('hide');
+      popup.innerHTML = '';
+      });
+
+      // comments section
+      let commentSection = `<div class="comments">
+        <h2 class="comments-heading">Comments</h2>
+        <h2 class="comments-num">(<span class="comments-count">0</span>)</h2>
+      </div>`;
+      const commentsArray = await getComment(ebook.trackId);
+      commentSection += `<div class="comment-list">${showComments(commentsArray)}</div>
+      <div class="add-comment">
+      <h2 class="text-center">Add a comment</h2>
+      <form action="" id="comment-form">
+        <input type="text" id="user-name" placeholder="Your name" required>
+        <textarea name="comment" id="user-comment" cols="30" rows="10" placeholder="Your insights" required></textarea>
+        <button type="submit" class="btn-comment">Comment</button>
+      </form>
+      </div>`;
+      popup.insertAdjacentHTML('beforeend', commentSection);
+
+      const commentList = document.querySelector('.comment-list');
+      const counterElement = document.querySelector('.comments-count');
+      counterElement.innerHTML = commentsCounter();
+
+      const userName = document.getElementById('user-name');
+      const userComment = document.getElementById('user-comment');
+      const form = document.getElementById('comment-form');
+
+      form.addEventListener('submit', async (e) => {
+        e.preventDefault();
+        await postComment(ebook.trackId, userName.value, userComment.value);
+        const commentsArray = await getComment(ebook.trackId);
+        commentList.innerHTML = showComments(commentsArray);
+        counterElement.innerHTML = commentsCounter();
+        userName.value = '';
+        userComment.value = '';
+      });
     });
     commentButton.textContent = 'Comments';
     commentButtonDiv.appendChild(commentButton);
   });
+};
+
+const showComments = (commentData) => {
+  let comments = '';
+  commentData.forEach((element) => {
+    comments += `<div class="comment-container">
+    <p class="comment-date">${element.creation_date}</p>
+    <p class="comment-name">${element.username}: </p>
+    <p class="comment-content">${element.comment}</p>
+  </div>`;
+  });
+  return comments;
 };
 
 export default renderMainList;
